AccountsServer : File Structure Changes ( types relocation and utils relocation )
