+++ packages/server/src/AccountsServer.ts
@@ -16,26 +16,18 @@ import {
   generateAccessToken,
   generateRefreshToken,
   generateRandomToken,
-} from './tokens';
-import { emailTemplates, EmailTemplateType, sendMail } from './email';
-import {
-  AccountsServerOptions,
-  ConnectionInformationsType,
-  AuthService,
-  DBInterface,
-} from './types';
+} from './utils/tokens';
 
-export interface TokenRecord {
-  token: string;
-  address: string;
-  when: number;
-  reason: string;
-}
+import { emailTemplates, sendMail } from './utils/email';
 
-export interface JwtData {
-  token: string;
-  isImpersonated: boolean;
-}
+import { AccountsServerOptions } from './types/AccountsServerOptions'
+import { ConnectionInformationsType } from './types/ConnectionInformationsType'
+import { AuthService } from './types/AuthService'
+import { DBInterface } from './types/DBInterface'
+import { TokenRecord } from './types/TokenRecord'
+import { JwtData } from './types/JwtData'
+import { RemoveListenerHandle } from './types/RemoveListenerHandle'
+import { EmailTemplateType } from './types/EmailTemplateType'
 
 const defaultOptions = {
   tokenSecret: 'secret',
@@ -52,8 +44,6 @@ const defaultOptions = {
   sendMail,
   siteUrl: 'http://localhost:3000',
 };
-
-export type RemoveListnerHandle = () => EventEmitter;
 
 export const ServerHooks = {
   LoginSuccess: 'LoginSuccess',
@@ -105,51 +95,51 @@ export class AccountsServer {
     return this.options;
   }
 
-  public onLoginSuccess(callback: HookListener): RemoveListnerHandle {
+  public onLoginSuccess(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.LoginSuccess, callback);
   }
 
-  public onLoginError(callback: HookListener): RemoveListnerHandle {
+  public onLoginError(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.LoginError, callback);
   }
 
-  public onLogoutSuccess(callback: HookListener): RemoveListnerHandle {
+  public onLogoutSuccess(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.LogoutSuccess, callback);
   }
 
-  public onLogoutError(callback: HookListener): RemoveListnerHandle {
+  public onLogoutError(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.LogoutError, callback);
   }
 
-  public onCreateUserSuccess(callback: HookListener): RemoveListnerHandle {
+  public onCreateUserSuccess(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.CreateUserSuccess, callback);
   }
 
-  public onCreateUserError(callback: HookListener): RemoveListnerHandle {
+  public onCreateUserError(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.CreateUserError, callback);
   }
 
-  public onResumeSessionSuccess(callback: HookListener): RemoveListnerHandle {
+  public onResumeSessionSuccess(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.ResumeSessionSuccess, callback);
   }
 
-  public onResumeSessionError(callback: HookListener): RemoveListnerHandle {
+  public onResumeSessionError(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.ResumeSessionError, callback);
   }
 
-  public onRefreshTokensSuccess(callback: HookListener): RemoveListnerHandle {
+  public onRefreshTokensSuccess(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.RefreshTokensSuccess, callback);
   }
 
-  public onRefreshTokensError(callback: HookListener): RemoveListnerHandle {
+  public onRefreshTokensError(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.RefreshTokensError, callback);
   }
 
-  public onImpersonationSuccess(callback: HookListener): RemoveListnerHandle {
+  public onImpersonationSuccess(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.ImpersonationSuccess, callback);
   }
 
-  public onImpersonationError(callback: HookListener): RemoveListnerHandle {
+  public onImpersonationError(callback: HookListener): RemoveListenerHandle {
     return this.on(ServerHooks.ImpersonationError, callback);
   }
 
@@ -542,7 +532,7 @@ export class AccountsServer {
     return this.db.setProfile(userId, { ...user.profile, ...profile });
   }
 
-  public on(eventName: string, callback: HookListener): RemoveListnerHandle {
+  public on(eventName: string, callback: HookListener): RemoveListenerHandle {
     this.hooks.on(eventName, callback);
 
     return () => this.hooks.removeListener(eventName, callback);
