+++ packages/server/src/index.ts
@@ -1,7 +1,11 @@
-import { AccountsServer } from './accounts-server';
-import * as encryption from './encryption';
-import { generateRandomToken } from './tokens';
-import { AuthService, DBInterface, ConnectionInformationsType } from './types';
+import { AccountsServer } from './AccountsServer';
+import * as encryption from './utils/encryption';
+import { generateRandomToken } from './utils/tokens';
+import { getFirstUserEmail } from './utils/getFirstUserEmail';
+
+import { ConnectionInformationsType } from './types/ConnectionInformationsType'
+import { AuthService } from './types/AuthService'
+import { DBInterface } from './types/DBInterface'
 
 export default AccountsServer;
 export {
@@ -11,4 +15,5 @@ export {
   DBInterface,
   generateRandomToken,
   ConnectionInformationsType,
+  getFirstUserEmail
 };
