+++ packages/server/src/types/ConnectionInformationsType.ts
@@ -1,0 +1,4 @@
+export interface ConnectionInformationsType {
+  ip?: string;
+  userAgent?: string;
+}
