+++ packages/server/src/types/DBInterface.ts
@@ -1,66 +1,5 @@
-import { UserObjectType, CreateUserType, SessionType } from '@accounts/common';
-import { EmailTemplatesType, EmailTemplateType, SendMailType } from './email';
-import { AccountsServer } from './accounts-server';
-
-export interface AuthService {
-  server: AccountsServer;
-  serviceName: string;
-  setStore(store: DBInterface): void;
-  authenticate(params: any): Promise<UserObjectType | null>;
-}
-
-export type UserObjectSanitizerFunction = (
-  userObject: UserObjectType,
-  omitFunction: (userDoc: object, fields: string[]) => UserObjectType,
-  pickFunction: (userDoc: object, fields: string[]) => UserObjectType
-) => any;
-
-export type ResumeSessionValidator = (
-  user: UserObjectType,
-  session: SessionType
-) => Promise<any>;
-
-export type PrepareMailFunction = (
-  to: string,
-  token: string,
-  user: UserObjectType,
-  pathFragment: string,
-  emailTemplate: EmailTemplateType,
-  from: string
-) => object;
-
-export type EmailType = EmailTemplateType & { to: string };
-
-export interface AccountsServerOptions {
-  db: DBInterface;
-  tokenSecret: string;
-  tokenConfigs?: {
-    accessToken?: {
-      expiresIn?: string;
-    };
-    refreshToken?: {
-      expiresIn?: string;
-    };
-  };
-  emailTokensExpiry?: number;
-  emailTemplates?: EmailTemplatesType;
-  userObjectSanitizer?: UserObjectSanitizerFunction;
-  impersonationAuthorize?: (
-    user: UserObjectType,
-    impersonateToUser: UserObjectType
-  ) => Promise<any>;
-  resumeSessionValidator?: ResumeSessionValidator;
-  siteUrl?: string;
-  prepareMail?: PrepareMailFunction;
-  sendMail?: SendMailType;
-  // https://github.com/eleith/emailjs#emailserverconnectoptions
-  email?: object;
-}
-
-export interface ConnectionInformationsType {
-  ip?: string;
-  userAgent?: string;
-}
+import { UserObjectType, SessionType, CreateUserType } from '@accounts/common';
+import { ConnectionInformationsType } from './ConnectionInformationsType';
 
 export interface DBInterface {
   // Find user by identity fields
