+++ packages/server/src/types/EmailTemplateType.ts
@@ -1,0 +1,7 @@
+import { UserObjectType } from '@accounts/common';
+ 
+export interface EmailTemplateType {
+  from?: string;
+  subject: (user?: UserObjectType) => string;
+  text: (user: UserObjectType, url: string) => string;
+}
