+++ packages/server/src/types/EmailTemplatesType.ts
@@ -1,0 +1,8 @@
+import { EmailTemplateType } from './EmailTemplateType'
+
+export interface EmailTemplatesType {
+  from: string;
+  verifyEmail: EmailTemplateType;
+  resetPassword: EmailTemplateType;
+  enrollAccount: EmailTemplateType;
+}
