+++ packages/server/src/types/EmailType.ts
@@ -1,0 +1,3 @@
+import { EmailTemplateType } from './EmailTemplateType'
+
+export type EmailType = EmailTemplateType & { to: string };
