+++ packages/server/src/types/JwtData.ts
@@ -1,0 +1,4 @@
+export interface JwtData {
+  token: string;
+  isImpersonated: boolean;
+}
