+++ packages/server/src/types/RemoveListenerHandle.ts
@@ -1,0 +1,2 @@
+import { EventEmitter } from 'events';
+export type RemoveListenerHandle = () => EventEmitter;
