+++ packages/server/src/types/ResumeSessionValidator.ts
@@ -1,0 +1,6 @@
+import { UserObjectType, SessionType } from '@accounts/common';
+
+export type ResumeSessionValidator = (
+  user: UserObjectType,
+  session: SessionType
+) => Promise<any>;
