+++ packages/server/src/types/SendMailType.ts
@@ -1,0 +1,1 @@
+export type SendMailType = (mail: object) => Promise<void>;
