+++ packages/server/src/types/TokenRecord.ts
@@ -1,0 +1,6 @@
+export interface TokenRecord {
+  token: string;
+  address: string;
+  when: number;
+  reason: string;
+}
