+++ packages/server/src/types/UserObjectSanitizerFunction.ts
@@ -1,0 +1,7 @@
+import { UserObjectType } from '@accounts/common';
+
+export type UserObjectSanitizerFunction = (
+  userObject: UserObjectType,
+  omitFunction: (userDoc: object, fields: string[]) => UserObjectType,
+  pickFunction: (userDoc: object, fields: string[]) => UserObjectType
+) => any;
