+++ packages/server/src/utils/getFirstUserEmail.ts
