+++ packages/server/src/utils/serverHooks.ts
@@ -1,0 +1,14 @@
+export const ServerHooks = {
+  LoginSuccess: 'LoginSuccess',
+  LoginError: 'LoginError',
+  LogoutSuccess: 'LogoutSuccess',
+  LogoutError: 'LogoutError',
+  CreateUserSuccess: 'CreateUserSuccess',
+  CreateUserError: 'CreateUserError',
+  ResumeSessionSuccess: 'ResumeSessionSuccess',
+  ResumeSessionError: 'ResumeSessionError',
+  RefreshTokensSuccess: 'RefreshTokensSuccess',
+  RefreshTokensError: 'RefreshTokensError',
+  ImpersonationSuccess: 'ImpersonationSuccess',
+  ImpersonationError: 'ImpersonationError',
+};
